pyproject.toml
[tool.poetry]
name = "com-analysis-backend"
version = "0.1.0"
description = ""
authors = ["kassraniroumand <kassra.niroumand@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.2"
python-multipart = "^0.0.18"
pydantic = {extras = ["email"], version = "^2.0"}
boto3 = "^1.35"
langchain = "^0.3"
langgraph = "^0.4"
langchain-aws = "^0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

